+++ azext_iot/central/models/ga_2022_07_31/organization.py
@@ -4,9 +4,8 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-
-class DeviceGroup:
-    def __init__(self, group: dict):
-        self.display_name = group.get("displayName")
-        self.id = group.get("id")
-        self.organizations = group.get("organizations")
+class Organization:
+    def __init__(self, org: dict):
+        self.display_name = org.get("displayName")
+        self.id = org.get("id")
+        self.parent = org.get("parent")