+++ HISTORY.rst
@@ -3,6 +3,34 @@
 Release History
 ===============
 
+unreleased
++++++++++++++++
+
+**IoT Central updates**
+
+* `--api-version` parameter will be deprecated and ignored. The IoT Central API will alway call latest GA version or latest preview version (if any API only exists in preview).
+
+* Add support for enrollment groups CRUD.
+
+  - az iot central enrollment-group
+    - az iot central enrollment-group list
+    - az iot central enrollment-group show
+    - az iot central enrollment-group create
+    - az iot central enrollment-group delete
+    - az iot central enrollment-group update
+    - az iot central enrollment-group verify-certificate
+    - az iot central enrollment-group generate-verification-code
+
+* Add support for scheduled jobs CRUD.
+
+  - az iot central scheduled-job
+    - az iot central scheduled-job list
+    - az iot central scheduled-job show
+    - az iot central scheduled-job create
+    - az iot central scheduled-job delete
+    - az iot central scheduled-job update
+    - az iot central scheduled-job list-runs
+
 0.17.2
 +++++++++++++++